Add auth but disable it for now
This commit documents how to add auth using Clerk. It also styles the
site a bit more and adds the ROS path planning documentation. This was
made possible thanks to Google Docs adding a markdown export feature
, so we can essentially just copy paste our entire google docs library
into the documentation site. This will probably be the final public
commit to the repo before I move it all onto a server.