Add some guides / tutorials / faq like material
I added a tonne of yap just to focus on content instead of style
briefly. I think part of any article which could be considered an
answer to an FAQ question ought to be mirrored to the team stack
overflow and also to an FAQ page on the site. We need to have a
defined "documentation improvement process" in which all documentation
is symbiotic e.g.

- Stackoverflow content bleeds into the site and the FAQ pages
- FAQ pages bleed into stackoverflow and the articles refer to
  the FAQ pages.
- Articles on the site become answers and refinements on existing
  FAQ pages